// src/app/interfaces/empleado.interface.ts
export interface Empleado{
    id: number;
    dni: string;
    apePaterno: string;
    apeMaterno: string;
    contrasenia: string;
    direccion: string;
    email: string;
    nacionalidad: string;
    nomUsuario: string;
    nombre:string;
    numTelefono: string;
    sexo:string;
    tipDocIdentificacion: string;
    area: string;
}